+++ packages/thumbprint-react/components/DatePicker/utilities.ts
@@ -8,28 +8,31 @@ import endOfDay from 'date-fns/end_of_day';
 import startOfDay from 'date-fns/start_of_day';
 import parse from 'date-fns/parse';
 
+import { DateIsh, PropTypes } from './types';
+
 // Throws an error `message` with a prefix showing that it comes from the DatePicker.
-function throwError(message) {
+function throwError(message: string): void {
     throw new Error(`TUI DatePicker: ${message}`);
 }
 
 // Returns true any of the given `dates` fall on a day before the day of `cutoff`.
-export function hasAnyPastDays(dates, cutoff = new Date()) {
+export function hasAnyPastDays(dates: Date[], cutoff: Date = new Date()): boolean {
     return some(dates, date => isBefore(endOfDay(date), cutoff));
 }
 
 // Returns true any of the given `dates` fall on a day after the day of `cutoff`.
-export function hasAnyFutureDays(dates, cutoff = new Date()) {
+export function hasAnyFutureDays(dates: Date[], cutoff: Date = new Date()): boolean {
     return some(dates, date => isAfter(startOfDay(date), cutoff));
 }
 
 // Normalise the `value` prop to always be an array of dates.
-export function normaliseValue(value) {
-    return map(castArray(value), parse);
+export function normaliseValue(value: PropTypes['value']): Date[] {
+    const valueArr: DateIsh[] = castArray<DateIsh>(value);
+    return map<DateIsh, Date>(valueArr, d => parse(d));
 }
 
 // Validate the `props` object and throw errors if there are any inconsistencies.
-export function validateProps(props) {
+export function validateProps(props: PropTypes): void {
     const days = normaliseValue(props.value);
 
     if (!props.allowMultiSelection && days.length > 1) {