+++ packages/thumbprint-react/components/Popover/index.tsx
@@ -1,47 +1,103 @@
 import React, { useState } from 'react';
-import PropTypes from 'prop-types';
 import classNames from 'classnames';
-import { Manager, Reference, Popper } from 'react-popper';
+import { Manager, Reference, Popper, RefHandler } from 'react-popper';
 import startsWith from 'lodash/startsWith';
 
 import * as tokens from '@thumbtack/thumbprint-tokens';
 
-import ConditionalPortal from '../../utils/ConditionalPortal.jsx';
+import ConditionalPortal from '../../utils/ConditionalPortal';
 import canUseDOM from '../../utils/can-use-dom';
 import useCloseOnEscape from '../../utils/use-close-on-escape';
 import useFocusTrap from '../../utils/use-focus-trap';
 
-import { TextButton } from '../Button/index.jsx';
-import { Themed } from '../UIAction/index.jsx';
+import { TextButton as _TextButton } from '../Button/index';
+import { Themed as _Themed } from '../UIAction/index';
 import { NavigationCloseTiny } from '../../icons/index.jsx';
 
 import styles from './index.module.scss';
 
-const Popover = ({
+// TS is attempting to infer types from the React.forwardRef call inside `Button` and `UIAction.
+// However, because they do not have prop types in TypeScript yet, it is not working properly.
+// Disable checking for now, but remove this line when we rewrite these components in TS.
+// TODO(giles): rewrite `TextButton` and `Themed` and remove this line.
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+const TextButton = _TextButton as any;
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+const Themed = _Themed as any;
+
+interface PopoverPropTypes {
+    /**
+     * Contents for the Popover. Usually a `PopoverTitle`, `PopoverBody`, and `PopoverPrimaryButton`
+     */
+    children: React.ReactNode;
+    /**
+     * A function that renders JSX and receives an object with `ref`.
+     * All of these props must be added to the component within the render prop.
+     */
+    launcher: ({ ref }: { ref: RefHandler }) => React.ReactNode;
+    /**
+     * Position of popover relative to the launcher.
+     */
+    position?:
+        | 'top-start'
+        | 'top'
+        | 'top-end'
+        | 'bottom-start'
+        | 'bottom'
+        | 'bottom-end'
+        | 'left-start'
+        | 'left'
+        | 'left-end'
+        | 'right-start'
+        | 'right'
+        | 'right-end';
+    /**
+     * Whether or not the popover is visible.
+     */
+    isOpen?: boolean;
+    /**
+     * Function called when the close button is clicked. You should cause this to set `isOpen=false`
+     * in your parent component.
+     */
+    onCloseClick: () => void;
+    /**
+     * By default popovers will render right before the `</body>` tag.
+     * Setting the `container` to `inline` causes the tooltip to remain where it was added to the
+     * DOM.
+     * This option is helpful to work around z-index and positioning issues.
+     */
+    container?: 'inline' | 'body';
+    /**
+     * Accessibility title used to describe the content of the popover to screen readers.
+     */
+    accessibilityLabel?: string;
+}
+
+export default function Popover({
     children,
     launcher,
-    position,
-    isOpen,
     onCloseClick,
-    container,
-    accessibilityLabel,
-}) => {
+    position = 'top',
+    isOpen = false,
+    container = 'body',
+    accessibilityLabel = 'Popover',
+}: PopoverPropTypes): JSX.Element {
     // Appends the tooltip right before `</body>` when true. Used to prevent z-index and positioning
     // issues.
     const shouldDisplace = container === 'body';
 
     // Using `useState` instead of `useRef `to allow multiple refs. See Image for another example.
-    const [wrapperEl, setWrapperEl] = useState(null);
+    const [wrapperEl, setWrapperEl] = useState();
 
-    const shouldTrapFocus = canUseDOM && wrapperEl;
-    const shouldBindEscListener = canUseDOM && isOpen;
+    const shouldTrapFocus: boolean = canUseDOM && !!wrapperEl;
+    const shouldBindEscListener: boolean = canUseDOM && isOpen;
 
     useCloseOnEscape(onCloseClick, shouldBindEscListener);
     useFocusTrap(wrapperEl, shouldTrapFocus);
 
     return (
         <Manager>
-            <Reference>{({ ref }) => launcher({ ref })}</Reference>
+            <Reference>{({ ref }): React.ReactNode => launcher({ ref })}</Reference>
             <ConditionalPortal shouldDisplace={shouldDisplace}>
                 {canUseDOM && (
                     <Popper
@@ -52,14 +108,14 @@ const Popover = ({
                         }}
                         positionFixed={false}
                     >
-                        {({ ref: popperRef, style, placement, arrowProps }) => (
+                        {({ ref: popperRef, style, placement, arrowProps }): JSX.Element => (
                             // Use tabIndex="-1" to allow programmatic focus (as initialFocus node
                             // for focus-trap) but not be tabbable by user.
                             <div
                                 role="dialog"
                                 aria-label={accessibilityLabel}
-                                tabIndex="-1"
-                                ref={el => {
+                                tabIndex={-1}
+                                ref={(el: HTMLElement | null): void => {
                                     setWrapperEl(el);
                                     popperRef(el);
                                 }}
@@ -103,127 +159,73 @@ const Popover = ({
             </ConditionalPortal>
         </Manager>
     );
-};
+}
 
-Popover.propTypes = {
+interface PopoverTitlePropTypes {
     /**
-     * Contents for the Popover. Usually a `PopoverTitle`, `PopoverBody`, and `PopoverPrimaryButton`
+     * The contents of the title.
      */
-    children: PropTypes.node.isRequired,
-    /**
-     * A function that renders JSX and receives an object with `ref`.
-     * All of these props must be added to the component within the render prop.
-     */
-    launcher: PropTypes.func.isRequired,
-    /**
-     * Position of popover relative to the launcher.
-     */
-    position: PropTypes.oneOf([
-        'top-start',
-        'top',
-        'top-end',
-        'bottom-start',
-        'bottom',
-        'bottom-end',
-        'left-start',
-        'left',
-        'left-end',
-        'right-start',
-        'right',
-        'right-end',
-    ]),
-    /**
-     * Whether or not the popover is visible.
-     */
-    isOpen: PropTypes.bool,
-    /**
-     * Function called when the close button is clicked. You should cause this to set `isOpen=false`
-     * in your parent component.
-     */
-    onCloseClick: PropTypes.func.isRequired,
-    /**
-     * By default popovers will render right before the `</body>` tag.
-     * Setting the `container` to `inline` causes the tooltip to remain where it was added to the
-     * DOM.
-     * This option is helpful to work around z-index and positioning issues.
-     */
-    container: PropTypes.oneOf(['inline', 'body']),
-    /**
-     * Accessibility title used to describe the content of the popover to screen readers.
-     */
-    accessibilityLabel: PropTypes.string,
-};
+    children: React.ReactNode;
+}
 
-Popover.defaultProps = {
-    isOpen: false,
-    position: 'top',
-    container: 'body',
-    accessibilityLabel: 'Popover',
-};
+const PopoverTitle = ({ children }: PopoverTitlePropTypes): JSX.Element => (
+    <div className={styles.popoverTitle}>{children}</div>
+);
 
-export default Popover;
-
-// === Sub-components ===
-
-const PopoverTitle = ({ children }) => <div className={styles.popoverTitle}>{children}</div>;
-
-PopoverTitle.propTypes = {
+interface PopoverBodyPropTypes {
     /**
-     * A title for the popover.
+     * The contents of the body.
      */
-    children: PropTypes.string.isRequired,
-};
+    children: React.ReactNode;
+}
 
-const PopoverBody = ({ children }) => <div className={styles.popoverBody}>{children}</div>;
+const PopoverBody = ({ children }: PopoverBodyPropTypes): JSX.Element => (
+    <div className={styles.popoverBody}>{children}</div>
+);
 
-PopoverBody.propTypes = {
+interface PopoverPrimaryButtonPropTypes {
     /**
-     * The body description text for the popover.
+     * The text of the button.
      */
-    children: PropTypes.node.isRequired,
-};
+    children: string;
+    /**
+     * The function to call when the button is clicked.
+     */
+    onClick: () => void;
+}
 
-const PopoverPrimaryButton = ({ children, onClick }) => (
+const PopoverPrimaryButton = ({
+    children,
+    onClick,
+}: PopoverPrimaryButtonPropTypes): JSX.Element => (
     <Themed size="small" onClick={onClick} theme="popover-primary">
         {children}
     </Themed>
 );
 
-PopoverPrimaryButton.propTypes = {
+interface PopoverSecondaryButtonPropTypes {
     /**
-     * The text for this button.
+     * The text of the button.
      */
-    children: PropTypes.string.isRequired,
+    children: string;
     /**
-     * Function called when the button is clicked.
+     * The function to call when the button is clicked.
      */
-    onClick: PropTypes.func.isRequired,
-};
+    onClick?: () => void;
+    /**
+     * Link to visit when the button is clicked. It will be opened in a new tab.
+     */
+    to?: string;
+}
 
-const PopoverSecondaryButton = ({ children, onClick, to }) => (
+const PopoverSecondaryButton = ({
+    children,
+    onClick,
+    to,
+}: PopoverSecondaryButtonPropTypes): JSX.Element => (
     <Themed size="small" onClick={onClick} to={to} theme="popover-secondary" shouldOpenInNewTab>
         {children}
     </Themed>
 );
-
-PopoverSecondaryButton.propTypes = {
-    /**
-     * The text for this button.
-     */
-    children: PropTypes.string.isRequired,
-    /**
-     * Function called when the button is clicked, or link to visit.
-     */
-    onClick: PropTypes.func,
-    /**
-     * Link to visit when the button is clicked. It will be opened in a new tab.
-     */
-    to: PropTypes.string,
-};
-
-PopoverSecondaryButton.defaultProps = {
-    onClick: undefined,
-    to: undefined,
-};
 
 export { PopoverTitle, PopoverBody, PopoverPrimaryButton, PopoverSecondaryButton };